+++ Glass Heart/Glass Heart/Source/Mode/ModeMain.h
@@ -7,6 +7,7 @@
  *********************************************************************/
 #pragma once
 #include <AppFrame.h>
+#include <cstdint>
  /**
 	* @namespace GlassHeart
 	* @brief プロジェクト名
@@ -83,10 +84,10 @@ namespace GlassHeart {
 
 			UI::UI& GetUI();
 		protected:
-			Application::GameMain& _gameMain; //!< ゲームメインクラスへの参照
-			std::int_fast8_t _fadeCount;      //!< 自動フェードインアウト用カウンタ
-			static std::int_fast8_t _count;                //!< タイマー用フレームカウント
-			static std::int_fast16_t _countSeconds;         //!< タイマー用フレームカウント
+			Application::GameMain& _gameMain;        //!< ゲームメインクラスへの参照
+			std::int_fast8_t _fadeCount;             //!< 自動フェードインアウト用カウンタ
+			static std::int_fast16_t _count;         //!< タイマー用フレームカウント
+			static std::int_fast16_t _countSeconds;  //!< タイマー用フレームカウント
 		};
 	} // namespace Mode
 } // namespace GlassHeart
